+++ src/Quarry.Benchmarks/Benchmarks/CteMultiBenchmarks.cs
@@ -9,8 +9,6 @@ using SqlKata.Compilers;
 
 namespace Quarry.Benchmarks.Benchmarks;
 
-// Reader floor here is bounded by Microsoft.Data.Sqlite.GetDecimal (string-parse path).
-// See CteSimpleBenchmarks.cs for the full explanation of why Dapper appears faster.
 public class CteMultiBenchmarks : BenchmarkBase
 {
     private const string MultiCteSql = """
@@ -37,7 +35,7 @@ public class CteMultiBenchmarks : BenchmarkBase
             results.Add(new OrderIdTotalDto
             {
                 OrderId = reader.GetInt32(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
@@ -84,7 +82,7 @@ public class CteMultiBenchmarks : BenchmarkBase
             results.Add(new OrderIdTotalDto
             {
                 OrderId = reader.GetInt32(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
