+++ src/Quarry.Benchmarks/Benchmarks/CteProjectionBenchmarks.cs
@@ -9,8 +9,6 @@ using SqlKata.Compilers;
 
 namespace Quarry.Benchmarks.Benchmarks;
 
-// Reader floor here is bounded by Microsoft.Data.Sqlite.GetDecimal (string-parse path).
-// See CteSimpleBenchmarks.cs for the full explanation of why Dapper appears faster.
 public class CteProjectionBenchmarks : BenchmarkBase
 {
     private const string CteProjectionSql = """
@@ -33,7 +31,7 @@ public class CteProjectionBenchmarks : BenchmarkBase
             results.Add(new OrderIdTotalDto
             {
                 OrderId = reader.GetInt32(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
@@ -86,7 +84,7 @@ public class CteProjectionBenchmarks : BenchmarkBase
             results.Add(new OrderIdTotalDto
             {
                 OrderId = reader.GetInt32(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
